rename example package names
