+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.mge_menstracker">
+    package="com.main.tracker">
 
     <application
         android:allowBackup="true"
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.NewCycleActivity" />
-        <activity android:name=".activities.NextExpectedCycleActivity" />
-        <activity android:name=".activities.MainActivity">
+        <activity android:name="com.main.tracker.activities.NewCycleActivity" />
+        <activity android:name="com.main.tracker.activities.NextExpectedCycleActivity" />
+        <activity android:name="com.main.tracker.activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
