+++ app/src/main/java/com/main/tracker/activities/MainActivity.kt
@@ -1,10 +1,10 @@
-package com.example.mge_menstracker.activities
+package com.main.tracker.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
-import com.example.mge_menstracker.R
+import com.main.tracker.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
