+++ app/src/main/java/com/main/tracker/activities/NextExpectedCycleActivity.kt
@@ -1,12 +1,12 @@
-package com.example.mge_menstracker.activities
+package com.main.tracker.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
-import com.example.mge_menstracker.calendars.DayViewContainer
-import com.example.mge_menstracker.R
+import com.main.tracker.calendars.DayViewContainer
+import com.main.tracker.R
 import com.kizitonwose.calendarview.CalendarView
 import com.kizitonwose.calendarview.model.CalendarDay
 import com.kizitonwose.calendarview.ui.DayBinder
